fix: Minor issues
