+++ src/main/java/it/gov/pagopa/pu/sil/exception/transcoder/handler/ConstraintViolationExceptionMessageTranscoder.java
@@ -29,7 +29,7 @@ public class ConstraintViolationExceptionMessageTranscoder implements ExceptionM
       .collect(Collectors.joining(";"));
 
     return new ExceptionMessageTranscoded(
-      PuSilErrorDTO.CategoryEnum.BAD_REQUEST.name(),
+      PuSilErrorDTO.CategoryEnum.BAD_REQUEST.getValue(),
       "Invalid request content." + errorDescription,
       errorFields
     );
