+++ src/test/java/it/gov/pagopa/pu/sil/config/rest/HttpClientErrorJsonBodyHandlerTest.java
@@ -1,10 +1,9 @@
 package it.gov.pagopa.pu.sil.config.rest;
 
-import it.gov.pagopa.pu.debtpositions.dto.generated.CategoryEnum;
-import it.gov.pagopa.pu.debtpositions.dto.generated.DebtPositionErrorDTO;
-import it.gov.pagopa.pu.debtpositions.dto.generated.ErrorFieldDTO;
 import it.gov.pagopa.pu.sil.config.json.JsonConfig;
-import it.gov.pagopa.pu.sil.connector.debtpositions.mapper.DebtPositionErrorDTOMapper;
+import it.gov.pagopa.pu.sil.dto.generated.ErrorFieldDTO;
+import it.gov.pagopa.pu.sil.dto.generated.PuSilErrorDTO;
+import it.gov.pagopa.pu.sil.dto.generated.PuSilErrorDTO.CategoryEnum;
 import it.gov.pagopa.pu.sil.exception.common.*;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -31,19 +30,19 @@ class HttpClientErrorJsonBodyHandlerTest {
   HttpClientErrorJsonBodyHandlerTest() throws URISyntaxException {
   }
 
-  private HttpClientErrorJsonBodyHandler<DebtPositionErrorDTO> buildHttpClientErrorHandler(boolean bodyPrinterWhenError) {
+  private HttpClientErrorJsonBodyHandler<PuSilErrorDTO> buildHttpClientErrorHandler(boolean bodyPrinterWhenError) {
     return new HttpClientErrorJsonBodyHandler<>(jsonMapper, "APPNAME", bodyPrinterWhenError,
-      DebtPositionErrorDTO.class, DebtPositionErrorDTOMapper::map);
+      PuSilErrorDTO.class, e -> new PuErrorDTO(e.getCategory().getValue(), e.getCode(), e.getMessage(), e.getFields()));
   }
 
   private final URI url = new URI("http://www.sample.com");
-  private final DebtPositionErrorDTO expectedErrorDTO = new DebtPositionErrorDTO(CategoryEnum.DEBT_POSITION_BAD_REQUEST, "BADREQUEST", "MESSAGE", List.of(new ErrorFieldDTO("FIELD", "FIELDERRORCODE", "FIELDERRORMESSAGE")), "TRACEID");
+  private final PuSilErrorDTO expectedErrorDTO = new PuSilErrorDTO(CategoryEnum.BAD_REQUEST, "BADREQUEST", "MESSAGE", List.of(new ErrorFieldDTO("FIELD", "FIELDERRORCODE", "FIELDERRORMESSAGE")), "TRACEID");
 
   @ParameterizedTest
   @ValueSource(booleans = {true, false})
   void testNo4xxException(boolean bodyPrinterWhenError) {
     // Given
-    HttpClientErrorJsonBodyHandler<DebtPositionErrorDTO> httpClientHandler = buildHttpClientErrorHandler(bodyPrinterWhenError);
+    HttpClientErrorJsonBodyHandler<PuSilErrorDTO> httpClientHandler = buildHttpClientErrorHandler(bodyPrinterWhenError);
     try (MockClientHttpResponse response = new MockClientHttpResponse(new byte[0], HttpStatus.SERVICE_UNAVAILABLE)) {
 
       // When
@@ -58,7 +57,7 @@ class HttpClientErrorJsonBodyHandlerTest {
   @ValueSource(booleans = {true, false})
   void testNoBodyException(boolean bodyPrinterWhenError) {
     // Given
-    HttpClientErrorJsonBodyHandler<DebtPositionErrorDTO> httpClientHandler = buildHttpClientErrorHandler(bodyPrinterWhenError);
+    HttpClientErrorJsonBodyHandler<PuSilErrorDTO> httpClientHandler = buildHttpClientErrorHandler(bodyPrinterWhenError);
     try (MockClientHttpResponse response = new MockClientHttpResponse(new byte[0], HttpStatus.BAD_REQUEST)) {
 
       // When
@@ -73,7 +72,7 @@ class HttpClientErrorJsonBodyHandlerTest {
   @ValueSource(booleans = {true, false})
   void testNotFoundException(boolean bodyPrinterWhenError) {
     // Given
-    HttpClientErrorJsonBodyHandler<DebtPositionErrorDTO> httpClientHandler = buildHttpClientErrorHandler(bodyPrinterWhenError);
+    HttpClientErrorJsonBodyHandler<PuSilErrorDTO> httpClientHandler = buildHttpClientErrorHandler(bodyPrinterWhenError);
     try (MockClientHttpResponse response = new MockClientHttpResponse(new byte[0], HttpStatus.NOT_FOUND)) {
 
       // When
@@ -88,7 +87,7 @@ class HttpClientErrorJsonBodyHandlerTest {
   @ValueSource(booleans = {true, false})
   void testBodyException(boolean bodyPrinterWhenError) {
     // Given
-    HttpClientErrorJsonBodyHandler<DebtPositionErrorDTO> httpClientHandler = buildHttpClientErrorHandler(bodyPrinterWhenError);
+    HttpClientErrorJsonBodyHandler<PuSilErrorDTO> httpClientHandler = buildHttpClientErrorHandler(bodyPrinterWhenError);
     try (MockClientHttpResponse response = new MockClientHttpResponse(jsonMapper.writeValueAsBytes(expectedErrorDTO), HttpStatus.BAD_REQUEST)) {
 
       // When
@@ -99,7 +98,7 @@ class HttpClientErrorJsonBodyHandlerTest {
       Assertions.assertEquals(expectedErrorDTO.getCategory().getValue(), result.getCategory());
       Assertions.assertEquals(expectedErrorDTO.getCode(), result.getCode());
       Assertions.assertEquals(expectedErrorDTO.getMessage(), result.getMessage());
-      Assertions.assertEquals(DebtPositionErrorDTOMapper.map(expectedErrorDTO).fields(), result.getFields());
+      Assertions.assertEquals(expectedErrorDTO.getFields(), result.getFields());
     }
   }
 
@@ -107,7 +106,7 @@ class HttpClientErrorJsonBodyHandlerTest {
   @ValueSource(booleans = {true, false})
   void testNoJsonBodyException(boolean bodyPrinterWhenError) {
     // Given
-    HttpClientErrorJsonBodyHandler<DebtPositionErrorDTO> httpClientHandler = buildHttpClientErrorHandler(bodyPrinterWhenError);
+    HttpClientErrorJsonBodyHandler<PuSilErrorDTO> httpClientHandler = buildHttpClientErrorHandler(bodyPrinterWhenError);
     try (MockClientHttpResponse response = new MockClientHttpResponse("INVALIDJSON".getBytes(), HttpStatus.BAD_REQUEST)) {
 
       // When
@@ -127,8 +126,8 @@ class HttpClientErrorJsonBodyHandlerTest {
 
   @Test
   void testBuildDefaultHttpClientExceptionTranscoder(){
-    BiFunction<HttpStatusCodeException, DebtPositionErrorDTO, RuntimeException> httpErrorTranscoder = HttpClientErrorJsonBodyHandler.buildDefaultHttpClientExceptionTranscoder("TEST", DebtPositionErrorDTO::getCode, DebtPositionErrorDTO::getMessage);
-    DebtPositionErrorDTO errorDTO = new DebtPositionErrorDTO(null, "BAD_REQUEST", "MESSAGE", null, null);
+    BiFunction<HttpStatusCodeException, PuSilErrorDTO, RuntimeException> httpErrorTranscoder = HttpClientErrorJsonBodyHandler.buildDefaultHttpClientExceptionTranscoder("TEST", PuSilErrorDTO::getCode, PuSilErrorDTO::getMessage);
+    PuSilErrorDTO errorDTO = new PuSilErrorDTO(null, "BAD_REQUEST", "MESSAGE", null, null);
 
     for (HttpStatus httpStatus : HttpStatus.values()) {
       RuntimeException result = httpErrorTranscoder
@@ -145,8 +144,8 @@ class HttpClientErrorJsonBodyHandlerTest {
 
   @Test
   void testBuildDefaultHttpClientExceptionTranscoder_noErrorCodeFunction(){
-    BiFunction<HttpStatusCodeException, DebtPositionErrorDTO, RuntimeException> httpErrorTranscoder = HttpClientErrorJsonBodyHandler.buildDefaultHttpClientExceptionTranscoder("TEST", null, DebtPositionErrorDTO::getMessage);
-    DebtPositionErrorDTO errorDTO = new DebtPositionErrorDTO(null, "BAD_REQUEST", "MESSAGE", null, null);
+    BiFunction<HttpStatusCodeException, PuSilErrorDTO, RuntimeException> httpErrorTranscoder = HttpClientErrorJsonBodyHandler.buildDefaultHttpClientExceptionTranscoder("TEST", null, PuSilErrorDTO::getMessage);
+    PuSilErrorDTO errorDTO = new PuSilErrorDTO(null, "BAD_REQUEST", "MESSAGE", null, null);
 
     for (HttpStatus httpStatus : HttpStatus.values()) {
       RuntimeException result = httpErrorTranscoder
